update date format in schedule-tasks endpoint
Updated the date format in the `schedule-tasks` endpoint to correctly parse the input. Instead of using separate `startOfDay` and `endOfDay` variables, now the `date` parameter is combined with the time from `startTime` and `endTime`. This change ensures that the correct date and time range is considered when scheduling tasks based on user availability.